Updated utility
Messed up my evaluation system but was able to fix it. Increased frontier for better options. Also added code to only add new nodes to be expanded if they have a better eval than the worst node. Beats Rand and SF 8/10 but in x2 the time. takes even longer with booger and wins 4/10